+++ R/.logo.R
@@ -3,14 +3,14 @@ library(hexSticker)
 # for windows
 sticker(
   "man/figures/moa.png",
-  s_x=1.4, s_y=0.85, s_width = 0.3,
-  package = "Elbird",
+  s_x = 1.4, s_y = 0.85, s_width = 0.3,
+  package = "elbird",
   p_size = 55, p_y = 1.3, p_x = 0.85,
   filename = "man/figures/logo.png",
   h_fill = "#A9C772",
   p_color = "#333333",
   h_color = "#73553D",
-  url = "mrchypark.github.io/Elbird",
+  url = "mrchypark.github.io/elbird",
   u_size = 9,
   u_color = "#333333",
   dpi = 500