+++ R/init.R
@@ -1,3 +1,4 @@
+
 init_chk_not <- function() {
   length(ls(envir=.el)) != 1
 }