+++ NEWS.md
@@ -1,3 +1,9 @@
+# elbird 0.0.2
+
+* package name to lower.
+* kiwi version 0.10.3 support.
+* python version 3.10.2 support.
+
 # Elbird 0.0.1
 
 * Fix add_user_word function with cache.